--- src/main/java/org/sert2521/bunnybots2024/commands/EmptyCommand.kt
@@ -1,4 +1,0 @@
-package org.sert2521.bunnybots2024.commands
-
-class EmptyCommand {
-}
